(#14852) Fix incorrect parsing of attributes after CN in the DN
Without this patch applied Puppet does not correctly parse out the
common name attribute from the distinguished name attribute of the
certificate presented by the agent.

This is a problem because some certificates may have attributes
following the CN field.  For example:

    subject= /C=UK/ST=Greater London/O=example/CN=mir.example.net/emailAddress=systems@example.net

Without this patch Puppet incorrectly parses the CN as
"mir.example.net/emailAddress=systems@example.net" when it should be
"mir.example.net"

This patch fixes the problem by changing the greedy "match anything"
regular expression to the greedy "match anything except the /
character."

This patch fixes the parsing problem for:

 * Mongrel rest
 * Mongrel xmlrpc
 * Rack rest
 * Rack xmlrpc