(#14843, #14844) Add block to Puppet::Provider.execpipe
Previously the execpipe wrapper in Puppet::Provider only accepted and passed a
block if called from a provider instance, and not the class. This led to
providers failiing with messages of "No block given". This commit brings the
class wrapper in line with the instance wrapper, adding a block to its argument
list and to its Puppet::Util::Execution.execpipe call.
It also adds a single test for the rpm package provider, to verify sanity of
self.instances.