(#14901) Use the last argument when there are duplicates
Previously, Trollop would raise an error if the command line contained
duplicate arguments. This caused the application to ignore all arguments
because of the error handling around Trollop. The argument handling will
now accept duplicate arguments, but will only use the last duplicate's
value. This is the same as the behaviour of 2.7.