Implement (most of) JRuby command execution support.
This is a basic implementation of the execution API for JRuby systems.  It is
complete except for support of UID and GID switching.

As written it allows us to get the test suite running at all, which is more
than we previously had.  UID and GID switching can wait until the rest of the
suite passes...

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>