(Maint) Limit data binding API information
Previously the data binding terminus took the entire scope object which
provided far to much access to the internals of puppet to the
implemetors of data binding terminuses. This changes it to pass through
only the data that should be needed for discovering what value to use in
most reasonable implementations of data binding.