Disable CA operation on JRuby.
The CRL implementation in JRuby 1.6.7, at least, causes a failure when
managing the CRL - it either adds a duplicate `crlNumber` extension, or the
provider does and we don't manipulate it as expected.

Either way this blocks correct operation as a CA on JRuby for the moment, so
we should honestly disable it and refuse to operate on an unsupported
platform.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>