Don't duplicate CHILD_STATUS: it can't be modified.
The object in $CHILD_STATUS can't be modified except by executing an external
command; it is literally impossible for anyone to change it in any other way.
(A fact that causes us pain in other parts of the code.)

Given that, duplicating it is useless in most places, rather than just
returning a reference to the same object.  The runtime also doesn't mutate in
place, it points the top level variable somewhere else.

In MRI the extraneous dup was just a waste of CPU cycles allocating the
object.  JRuby literally can't allocate a new Process::Status object in Ruby
code, and caused a hard failure in consequence.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>