(maint) Add symlink stub to gentoo service provider spec
The gentoo service provider spec test stubbing didn't stub symlinks. A recent
change in the init provider started discarding symlinks from upstart jobs and
because this element wasn't stubbed, the test went to the (ubuntu) filesystem
and found a symlink to an upstart job, which was not the intent of the test.
This commit adds a blanket stub that there are no symlinks in the filesystem
for the specific test.