Avoid dangling symlinks for the sake of JRuby
JRuby 1.6.7.2 (and 1.7.0-preview1) have a bug where they are unable to
`readlink` on a dangling symlink.  This isn't actually needed in this
particular test - it was just a happenstance of implementation.

So, if we create the second target file as well as the first we can have the
suite pass on both platforms before the upstream problem is resolved.

See https://github.com/jruby/jruby/issues/210 for the upstream bug report.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>