Rewrite binwrite emulation again, this time for JRuby.
This emulation is an ongoing source of pain: every obscure platform has a
different set of things that go wrong, or are not implemented, or are
not abstracted.

In this case the IO.new mode operation on JRuby doesn't support POSIX numeric
modes, only string modes.  That blows up the last way we tried to work around
this stuff.

On the plus side, now we have a slow, ugly, but portable implementation that
passes on Windows, JRuby, and MRI 1.8.  This is good, finally.

Writing portable Ruby is just like writing portable C, only slower.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>