Disable `safe_posix_fork` tests on JRuby.
JRuby doesn't implement fork, but does explode if you close off an FD that it
uses; other platforms have tests that stub out fork itself, but let the other
behaviour alone, and that work exactly and only by chance.

JRuby doesn't - but we can disable the tests, as it doesn't break anything
deeper in the system that wouldn't already be broken.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>