Skip ~ expansion failure test that passes on JRuby
JRuby actually goes out and does the right thing  when faced with expanding a
path containing a tilde, but when 'HOME' is not set in the environment.
This is unlike MRI, which just fails in that case.

This is actually desirable behaviour, so just skip the test that asserts we
warn in that failure case.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>