Avoid stubbing methods on `nil`
Recent versions of Mocha treat stubbing something on the `nil` object as an
error, which is reasonable, but we did that in the queue application spec.

Since it makes no great difference, we can just put a stub object in place
where nil was sitting, and let that save us the trouble instead.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>