(#14909) Update createpackage.sh to resolve permissions issues
Previously, PackageMaker would apply the permissions found on
the system to files in the package if one of the files in the
payload already existed on the disk, rather than the ones set
up in the package root.

This means that when the Mac Puppet package had previously installed
group-writable files, and PackageMaker was used to build a new
package, the files in the new package used the local broken
permissons - resulting in a package containing files with group-
writable permissions.

This commit changes the call to PackageMaker to explicitly not
apply recommended permissions to package contents (using the
'--no-recommend' flag) . This ensures that the desired permissions
are applied.