(#7316) Restrict use of libdir and vardir settings in config files
This pull request is a short-term fix for a configuration
problem that we have uncovered.  Because the agent syncs
down code during pluginsync that needs to be accessible
to all other puppet client programs, it is no longer
acceptable to allow different custom libdir settings for
the agent; if we allow this then the other client tools
will fail when they look in a different libdir for the
code.

Longer term we need to do a larger cleanup of the config files,
but for now, this patch simply disallows libdir and vardir
anywhere in the config file other than "main" and "master".