`expand_path` eliminates `..` directories under JRuby.
The generate function test used `File.expand_path` to fully qualify a command
name that included the `..` sequence, for Windows.  Unfortunately, this
actually washes out the sequence entirely on JRuby - leading to a test failure
because the command was pre-cleaned of the failing input.

We can limit the "make absolute" part to only Windows, in which case we get
the correct behaviour as well as more clarity on why it is done.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>