Fix bad stubbing in dpkg package provider spec.
The tests depended on stubbing the internal implementation of the Tempfile
class to work - they depended on `Tempfile.new` being invoked inside
`Tempfile.open` to behave correctly.

Unfortunately, JRuby doesn't expose that level of implementation detail in a
way that the stubbing works as intended - so we either need to rethink our
testing strategy, or stub the actual call the implementation makes.

This follows the later path: it is simpler, and no worse, than the previous
code was.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>