Use more realistic certificates for testing.
JRuby demanded that a genuine key be available on any X.509 certificate that
was used.  This revealed that we had a half completed issuer certificate in
our testing here - we didn't give it a key.

This creates the key, since we should test in as realistic an environment as
possible for code contained within our current object.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>