(#15346) add --source to the gem list command
Without this patch the `gem list` command in the gem package provider
will search all package repositories.  This is a problem because the end
user is likely expecting Puppet to only search the specified repository
instead of all repositories.

This patch fixes the problem by passing the source parameter value onto
the gem list command as the `--source` command line option.

    Commit message amended by Jeff McCune <jeff@puppetlabs.com>