Ensure ZAML strings are flagged as UTF-8 content.
When encoding ZAML content we need to be sure we flag the results as UTF-8 on
platforms that support string encoding.

This makes little difference when touching the outside world, most of the
time, but on the strictest platforms - like JRuby - it can mean the difference
between being able to read the YAML content back correctly or not.

This is also technically correct in conformance with the YAML specification.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>