Performance: TypeCollection should remember nonexistent types before autoloading
When the TypeCollection doesn't find a class/definition it tries to
autoload it in each namespace. This requires calling TypeLoader#import
for this qualified name, which is a time consuming operation.

When loading definitions/classes that can't be found by the autoloader
(ie legacy manifests for instance),  the autoloader will still try to
load them in the autoloaded module path, even though those types have
possibly already been loaded (ie directly imported).  Searching for
those definitions/classes then becomes really expensive.

This problem is frequent for people (like me) that migrated manifests
from the 0.24/0.25 era without refactoring all their modules to strictly
follow the autoloading mechanism.

Now, we negatively cache when we couldn't load a type with the
autoloader, so that the next attempt we won't spend our time trying
again.

This saves about 30% of compilation time on my setup

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>