(#15187) Ensure that fatal errors are logged to console
This commit ensures that the agent always opens a console
logging destination.  It will be closed when the agent
has successfully initialized itself and switched to
daemon mode, but if any errors occur before that point, the
log messages will be available on the console.  Prior to
this commit, fatal errors during startup were not logged
to the console, which made it seem like the error
was being swallowed or that the agent had started successfully
even when it had not.