(#15320) pmt generate generated _gitignore instead of .gitignore
So the file stored in the main skeleton was _gitignore. Previously the code
specified a rewrite from an underscore to a dot, but alas this did not work in
ruby 1.8.7.

I'm removing that code for now, and renaming the file to .gitignore. I can see
why people might want the rename behaviour, but this has no tests and doesn't
work on ruby 1.8.7. For now to get the gitignore working I've reverted this
behaviour introduced in 94808b89eb27d13c5910ac5c6704e056d6cd46d5.