(maint) Standardize on /usr/bin/env ruby -S rspec
Without this patch some spec files are using `ruby -S rspec` and others
are using `rspec`.

We should standardize on a single form of the interpreter used for spec
files.

`ruby -S rspec` is the best choice because it correctly informs editors
such as Vim with Syntastic that the file is a Ruby file rather than an
Rspec file.