Fix broken TransBucket transformation tests.
This fixes the membership test to check actual resources rather than checking
if the TransBucket object itself is part of the catalog - it isn't, and hasn't
been for approximately ever.

It also eliminates a test that resources are "finalized" - a concept that
isn't valid in the code and hasn't been for years, if ever.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>