zaml: don't use backreferences on strings
While the YAML spec requires that we emit a backreference if the same string
object occurs twice in our content, it also costs a lot.  About two seconds of
time encoding the output, in my testing, out of a total of three.

Everything we do internally with YAML doesn't actually *depend* on the ability
to serialize an object graph that treats String objects as duplicated.

This is reinforced by every object that can serialize to PSON - since
it *can't* have a back-reference, you automatically get string duplication
there.

That means that we can just cut those few seconds out of the code without any
particular cost or suffering, and shave *way* more time off a run than almost
anything else we can do to our YAML output code.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>