A node always has an environment.
Puppet node objects, now, always have environments, and they are the genuine
Ruby object representing an environment, never just the name.

In consequence of this, the code in the compiler to handle two conditions - no
environment, or an empty string as environment - that presumably could have
been true at some time in the past ... no longer makes sense.

Instead, replace it with a clear and direct failure in the event we don't get
the object we expect, and carry on.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>